==> Source Code/Crawling/1_Crawl_transfermarkt_teams.ipynb <==
import pandas as pd
import re
import numpy as np
from bs4 import BeautifulSoup
import souper_v2

# Danh sách các giải đấu
URLs_transfermkt = {
                 'Premier League (ENG)':'https://www.transfermarkt.co.uk/premier-league/startseite/wettbewerb/GB1',
                 'Bundesliga (GER)': 'https://www.transfermarkt.co.uk/bundesliga/startseite/wettbewerb/L1',
                 'La Liga (SPA)': 'https://www.transfermarkt.co.uk/laliga/startseite/wettbewerb/ES1',
                 'Serie A (ITA)': 'https://www.transfermarkt.co.uk/serie-a/startseite/wettbewerb/IT1',
                 'Ligue 1 (FRA)': 'https://www.transfermarkt.co.uk/ligue-1/startseite/wettbewerb/FR1',
                 'Eredivise (NET)': 'https://www.transfermarkt.co.uk/eredivisie/startseite/wettbewerb/NL1',
                 'Brasileiro (BRA)': 'https://www.transfermarkt.co.uk/campeonato-brasileiro-serie-a/startseite/wettbewerb/BRA1',
                 'Liga NOS (POR)': 'https://www.transfermarkt.co.uk/liga-nos/startseite/wettbewerb/PO1',
                 'Liga MX (MEX)': 'https://www.transfermarkt.co.uk/liga-mx-clausura/startseite/wettbewerb/MEX1',
                 'Premier Liga (RUS)': 'https://www.transfermarkt.co.uk/premier-liga/startseite/wettbewerb/RU1',
                 'Championship (ENG)': 'https://www.transfermarkt.co.uk/championship/startseite/wettbewerb/GB2',
                 'Super Lig (TUR)': 'https://www.transfermarkt.co.uk/super-lig/startseite/wettbewerb/TR1',
                 'Bundesliga (AUT)': 'https://www.transfermarkt.co.uk/bundesliga/startseite/wettbewerb/A1',
                 'MLS (USA)': 'https://www.transfermarkt.co.uk/major-league-soccer/startseite/wettbewerb/MLS1',
                 'Primera División (ARG)': 'https://www.transfermarkt.co.uk/primera-division/startseite/wettbewerb/AR1N',
                 'J-League (JPN)': 'https://www.transfermarkt.co.uk/j1-league/startseite/wettbewerb/JAP1',
                 'Saudi Pro League (KSA)': 'https://www.transfermarkt.co.uk/saudi-pro-league/startseite/wettbewerb/SA1',
                 'K-League 1 (KOR)': 'https://www.transfermarkt.co.uk/k-league-1/startseite/wettbewerb/RSK1',
                 'DStv Premiership (ZAF)': 'https://www.transfermarkt.co.uk/dstv-premiership/startseite/wettbewerb/SFA1',
                 'A-League (AUS)': 'https://www.transfermarkt.co.uk/a-league-men/startseite/wettbewerb/AUS1'
                 }

sp = souper_v2.souper()

def extract_teams_or_players(soup, league_or_team, team):
    
    returner_df = pd.DataFrame()
    links = []
    teams_or_players = []
    league_or_team_list = []
    
    if team:
        team_pattern = re.compile(r'(?<=\ alt=").+(?=" class)')
        link_pattern = re.compile(r'(?<=\ href=").+[startseite].+(?=/saison_id/)')

        for each in soup.find('table', class_ = "items").find_all('a'): #, attrs={'class':'vereinprofil_tooltip tooltipstered'}
            
            if 'startseite' in str(each):

                try:
                    team_or_player = re.findall(team_pattern, str(each))[0]
                    link = re.findall(link_pattern, str(each))[0]
                    team_or_player = team_or_player.strip()
                    link = link.strip()
                    teams_or_players.append(team_or_player)
                    links.append(link)
                    league_or_team_list.append(league_or_team)

                except:
                    None
        
    else:
        player_pattern = re.compile(r'(?<=\">).+(?=</a>)')
        link_pattern = re.compile(r'(?<=\ href=").+(?=" id=)')
        
        table = soup.find("table", class_ = "items")

        for each in table.find_all('a'): #, attrs={'class': 'spielprofil_tooltip tooltipstered'}

            try:
                team_or_player = re.findall(player_pattern, str(each))[0]
                link = re.findall(link_pattern, str(each))[0]
                team_or_player = team_or_player.strip()
                link = link.strip()
                teams_or_players.append(team_or_player)
                links.append(link)
                league_or_team_list.append(league_or_team)

            except:
                None
    
    if team:
        returner_df['league'] = league_or_team_list
        returner_df['teams'] = teams_or_players
    else:
        returner_df['team'] = league_or_team_list
        returner_df['player'] = teams_or_players
        
    returner_df['link'] = links
    returner_df = returner_df.drop_duplicates(subset='link', keep="last")

    return returner_df

def get_teams(URLs_dict, team=True):
    
    teams_detail_df = pd.DataFrame()
    
    for each in range(len(URLs_dict.keys())):
        soup = sp.get_soup(list(URLs_dict.values())[each])
        league = list(URLs_dict.keys())[each]
        teams = extract_teams_or_players(soup, league, team)
        print(teams)
        teams_detail_df = pd.concat([teams_detail_df, teams], axis=0, sort=True)
        
        print("Complete " + league)
        
    return teams_detail_df

teams_transfermkt = get_teams(URLs_transfermkt)

print(teams_transfermkt)

teams_transfermkt.to_csv(r'C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Obtaining_players_and_teams/teams_transfermarket.csv',index=False)

transfermkt_teams = pd.read_csv('C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Obtaining_players_and_teams/teams_transfermarket.csv')

tm = pd.DataFrame(columns = ['league', 'link', 'teams'])
for index, row in transfermkt_teams.iterrows():
    if ">" not in row['teams']:
        tm = tm.append(row, ignore_index=True)
print(len(list(tm['teams'])))

tm.to_csv(r'C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Obtaining_players_and_teams/teams_transfermarket.csv',index=False)

transfermkt_teams = pd.read_csv('C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Obtaining_players_and_teams/teams_transfermarket.csv')

li = list(transfermkt_teams['teams'])
print(li)

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

print(transfermkt_teams)
