==> Source Code/Crawling/2_Crawl_players.ipynb <==
import pandas as pd
import re
import numpy as np
from bs4 import BeautifulSoup
from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.support.ui import WebDriverWait
from selenium.common.exceptions import TimeoutException
from selenium.webdriver.common.by import By
from selenium.webdriver.support import expected_conditions as EC
import time

# Hàm thu thập danh sách các cầu thủ từ một đội bóng
def extract_players(soup, team):
    
    returner_df = pd.DataFrame(columns = ['team', 'player', 'link'])
    links = []
    players = []
    team_list = []
    
    player_pattern = re.compile(r'(?<=\">).+(?=</a>)')
    link_pattern = re.compile(r'(?<=\ href=").+(?=" id=)')
    

    for each1 in soup.find_all('div', class_ = 'di nowrap'): #, attrs={'class': 'spielprofil_tooltip tooltipstered'}
        
        each2 = each1.find('span', class_ = 'hide-for-small')
        
        if each2 is None:
            continue
        
        each = each2.find('a')
        
        if each is None:
            continue

        try:
            link = each['href']
            player = each.text
            players.append(player)
            links.append(link)
            team_list.append(team)
            '''
            player = re.findall(player_pattern, str(each))[0]
            link = re.findall(link_pattern, str(each))[0]
            player = player.strip()
            link = link.strip()
            players.append(player)
            links.append(link)
            team_list.append(team['teams'])
            '''

        except:
            None
    
    
    returner_df['team'] = team_list
    returner_df['player'] = players
        
    returner_df['link'] = links
    returner_df = returner_df.drop_duplicates(subset='link', keep="last")

    return returner_df

def get_soup(URL):
    
    chrome_options = Options()
    chrome_options.add_argument('--window-size=1,1')
    
    
    driver = webdriver.Chrome(ChromeDriverManager().install(), options = chrome_options)
    wait = WebDriverWait(driver, 10)
    driver.get(URL)
    if len(driver.find_elements_by_css_selector("[name=ROBOTS]")) > 0:
        driver.get(URL)
    soup = BeautifulSoup(driver.page_source)
    driver.close()
    return soup

# Hàm thu thập danh sách cầu thủ từ danh sách các đội bóng
def get_players(teams_df):
    
    players_detail_df = pd.DataFrame()
    
    for each in range(teams_df.shape[0]):
        
        URL = 'https://www.transfermarkt.co.uk' + teams_df.iloc[each]['link']
        print(URL)
        soup = get_soup(URL)
        players = extract_players(soup, teams_df.iloc[each]['teams'])
        print(players)
        
        players_detail_df = pd.concat([players_detail_df, players], axis=0, sort=True)
        print("Complete tm " + teams_df.iloc[each]['teams'])
        
    return players_detail_df

transfermkt_teams = pd.read_csv('C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Obtaining_players_and_teams/teams_transfermarket.csv')

for i in range(transfermkt_teams.shape[0]):
    print(transfermkt_teams.iloc[i]['teams'])

players = get_players(transfermkt_teams)

print(players)

players.to_csv('C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Obtaining_players_and_teams/players_transfermarket.csv')
