==> Source Code/Crawling/3_Crawl_basic_info_from_players.ipynb <==
import pandas as pd
import numpy as np
import time
import re
from tqdm import tqdm_notebook
from torrequest import TorRequest
from bs4 import BeautifulSoup
from retrying import retry
from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.support.ui import WebDriverWait
from selenium.common.exceptions import TimeoutException
from selenium.webdriver.common.by import By
from selenium.webdriver.support import expected_conditions as EC

# Danh sách các cầu thủ đã được thu thập trước đó
sampled_players = pd.read_csv('C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Obtaining_players_and_teams/players_transfermarket.csv')

def get_soup(URL):
    
    chrome_options = Options()
    chrome_options.add_argument('--window-size=1,1')
    
    #options = webdriver.ChromeOptions()
    #options.add_argument('--disable-popup-blocking')
    
    driver = webdriver.Chrome(ChromeDriverManager().install(), options = chrome_options)
    wait = WebDriverWait(driver, 10)
    driver.get(URL)
    if len(driver.find_elements_by_css_selector("[name=ROBOTS]")) > 0:
        driver.get(URL)
    soup = BeautifulSoup(driver.page_source)
    driver.close()
    return soup

print(sampled_players.shape)

def extract_player_info_block(players_df):
    
    #links = players_df.link_ws
    #teams = players_df.team
    
    player_info = {'name': [],
                  'position': [],
                  'height': [],
                  'age': [],
                  'team': []} 

    for index in range(players_df.shape[0]):
        URL = 'https://www.transfermarkt.co.uk' + players_df.iloc[index]['link']
        print(URL)
        soup = get_soup(URL)
        text = ""
        div = soup.find('div', class_ = 'info-table info-table--right-space')
        
        if div is None:
            #soup = get_soup(URL)
            div = soup.find('div', class_ = 'info-table info-table--right-space min-height-audio')
            
        key_list = div.find_all('span', class_ = 'info-table__content info-table__content--regular')
        value_list = div.find_all('span', class_ = 'info-table__content info-table__content--bold')
        
        
        # Get the values
        h = 0
        a = 0
        p = 0
        for i in range(len(key_list)):
            if key_list[i].text == "Height:":
                # Preprocess the height of the player
                height_str = value_list[i].text
                height_str = height_str.replace(',', '.')  # Replace comma with period for decimal separator
                height_str = height_str.replace('\xa0', '')  # Remove non-breaking space character
                height_str = height_str.replace('m', '')  # Remove the unit 'm'
                height_str = height_str.replace('،', '.')
                height = float(height_str)
                height = int(height * 100)  # Convert to centimeters
                player_info['height'].append(height)
                h = 1
            elif key_list[i].text == "Age:":
                age = int(value_list[i].text)
                player_info['age'].append(age)
                a = 1
            elif key_list[i].text == "Position:":
                # Preprocess the position of the player
                pos = value_list[i].text
                pos = pos.replace(' ', '')
                pos = pos[1:]
                if pos.count("-") >= 2:
                    pos = pos.replace("-", " ", 1)
                player_info['position'].append(pos)
                p = 1

        player_info['name'].append(players_df.iloc[index]['player'])
        player_info['team'].append(players_df.iloc[index]['team'])
        
        if h == 0:
            player_info['height'].append(None)
        if a == 0:
            player_info['age'].append(None)
        if p == 0:
            player_info['position'].append(None)

            
        print(index, "Complete", players_df.iloc[index]['player'])
        
        '''    
            try:
                text = text + str(each)
            except:
                None
        
        name = re.findall(r'((?<=Name:</dt>\n<dd>).+(?=</dd>\n</dl><dl class))', text)
        height = re.findall(r'((?<=Height:</dt>\n<dd>).+(?=cm</dd>))', text)
        position = re.findall(r'((?<=Positions:</dt>\n<dd>\n<ul>\n<li>).*(?= |</li>))', text)
        weight = re.findall(r'((?<=Weight:</dt>\n<dd>).+(?=kg))', text)
        
        player_info['name'].append(name[0]) if len(name) >0 else player_info['name'].append(np.nan)
        player_info['position'].append(position[0].strip()) if len(position) >0 else player_info['position'].append(np.nan)
        player_info['height'].append(height[0]) if len(height) >0 else player_info['height'].append(np.nan)
        player_info['weight'].append(weight[0]) if len(weight) >0 else player_info['weight'].append(np.nan)
        player_info['team'].append(each_team)
        '''


    player_info_df = pd.DataFrame.from_dict(player_info, orient='index')
    player_info_df = player_info_df.transpose()
    
    return player_info_df

# Vì danh sách cầu thủ nhiều, nếu cào hết 1 lần rất tốn tài nguyên 
# -> chia data thành các sample nhỏ để dễ dàng crawl thông tin của từng cầu thủ hơn
# Sample_data là một list, mỗi phần tử là 1 dataframe tách ra từ sampled_players
sample_data = []
sample_data.append(sampled_players.iloc[:10])
i = 10
while i < sampled_players.shape[0]:
    if i >= sampled_players.shape[0] - 200:
        sample_data.append(sampled_players.iloc[i:])
    else:
        sample_data.append(sampled_players.iloc[i:i+200])
    i += 200 # mỗi sample chứa 200 dòng
print(len(sample_data))

print(sample_data[48])

# Muốn cào sample nào thì thay i bàng sample đó
i = 39
sample = extract_player_info_block(sample_data[i])

print(sample)

# Lưu thành file sample_data_i.csv
file_name = "C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Basic_info_player/sample_data_" + str(i) + ".csv"
sample.to_csv(file_name)
