==> Source Code/Crawling/4_merge_basic_info_player.ipynb <==
import pandas as pd

players = pd.read_csv('C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Obtaining_players_and_teams/players_transfermarket.csv')
basic_info = pd.read_csv('C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Basic_info_player/basic_info.csv')

basic_info_with_link = pd.merge(players, basic_info, left_on=['player','team'], right_on=['name', 'team'])
basic_info_with_link

basic_info_with_link.to_csv('C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Basic_info_player/basic_info_with_link.csv')
