==> Source Code/Crawling/5_Crawl_players_stats.ipynb <==
import pandas as pd
import numpy as np
import time
import re
from tqdm import tqdm_notebook
from torrequest import TorRequest
from bs4 import BeautifulSoup
from retrying import retry
from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.support.ui import WebDriverWait
from selenium.common.exceptions import TimeoutException
from selenium.webdriver.common.by import By
from selenium.webdriver.support import expected_conditions as ECimport 
import pandas as pd
import souper_v2
from tqdm import tqdm_notebook

sampled_players = pd.read_csv('C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Basic_info_player/basic_info_with_link.csv')
sampled_players

# Backup of original simplier function to get_soup
def get_soup(URL):
    
    chrome_options = Options()
    chrome_options.add_argument('--window-size=1,1')
    
    #options = webdriver.ChromeOptions()
    #options.add_argument('--disable-popup-blocking')
    
    driver = webdriver.Chrome(ChromeDriverManager().install(), options = chrome_options)
    wait = WebDriverWait(driver, 10)
    driver.get(URL)
    if len(driver.find_elements_by_css_selector("[name=ROBOTS]")) > 0:
        driver.get(URL)
    soup = BeautifulSoup(driver.page_source)
    driver.close()
    return soup

# Thu thập dữ liệu chuyên môn của từng cầu thủ
def obtain_players_stats(players_df):
    
    columns = ['player', 'appearance', 'goals', 'assists', 'yellow cards',
               'second yellow cards', 'red cards', 'goals conceded', 'clean sheets', 'minutes played']
    
    stats_player_df = pd.DataFrame(columns = columns)

    for index in range(players_df.shape[0]):
        stats_each_df = pd.DataFrame(columns = columns)
        
        results = 0
        each_link = players_df.iloc[index]['link']
        
        new_link = each_link.replace("profil","leistungsdatendetails")
        new_link = 'https://www.transfermarkt.co.uk' + new_link
        print(new_link)
        
        soup = get_soup(new_link)
        table = soup.find('table', class_ = 'items')
        
        if table is None or new_link == 'https://www.transfermarkt.co.uk/mohammed-jaseem/leistungsdatendetails/spieler/686068':
            stats = {'player':[], 'appearance': [], 'goals': [], 'assists': [], 'yellow cards': [],
            'second yellow cards': [], 'red cards': [], 'goals conceded': [], 'clean sheets': [], 'minutes played': []}
            
            stats['player'].append(each_link)
            stats['appearance'].append(0)
            stats['goals'].append(0)
            stats['assists'].append(0)
            stats['goals conceded'].append(0)
            stats['clean sheets'].append(0)
            stats['yellow cards'].append(0)
            stats['second yellow cards'].append(0)
            stats['red cards'].append(0)
            stats['minutes played'].append(0)
            
            stats1 = pd.DataFrame(stats)
            
            stats_player_df = pd.concat([stats_player_df, stats1])
            
            continue
        
        body = table.find('tbody')
        
        for row in body.find_all('tr'):
            stats = {'player':[], 'season': [], 'appearance': [], 'goals': [], 'assists': [], 'yellow cards': [],
            'second yellow cards': [], 'red cards': [], 'goals conceded': [], 'clean sheets': [], 'minutes played': []}
            
            stats['player'].append(each_link)
            
            td = row.find_all('td')
            
            stats['season'].append(td[0].text)
            stats['appearance'].append(int(td[4].find('a').text.replace('-','0')))
            stats['goals'].append(int(td[5].text.replace('-','0')))
            card_index = 0
        
            if players_df.iloc[index]['position'] == 'Goalkeeper':
                stats['assists'].append(0)
                stats['goals conceded'].append(int(td[7].text.replace('-','0')))
                stats['clean sheets'].append(int(td[8].text.replace('-','0')))
                # Minutes played
                minutes = td[9].text.strip()
                minutes = minutes.replace("'",'')
                minutes = minutes.replace(".",'')
                minutes = minutes.replace('-','0')
                stats['minutes played'].append(int(minutes))
                card_index = 6
                
            else:
                stats['goals conceded'].append(0)
                stats['clean sheets'].append(0)
                stats['assists'].append(int(td[6].text.replace('-','0')))
                # Minutes played
                minutes = td[8].text.strip()
                minutes = minutes.replace("'",'')
                minutes = minutes.replace(".",'')
                minutes = minutes.replace('-','0')
                stats['minutes played'].append(int(minutes))
                card_index = 7
            
            card = td[card_index].text.strip()
            pattern = r"([^/]+)/([^/]+)/([^/]+)"
            matches = re.findall(pattern, card)
            for match in matches:
                a,b,c = match
                stats['yellow cards'].append(int(a.replace('-','0')))
                stats['second yellow cards'].append(int(b.replace('-','0')))
                stats['red cards'].append(int(c.replace('-','0')))
            
            stats_df = pd.DataFrame.from_dict(stats, orient='index')
            stats_df = stats_df.transpose()
            
            stats_each_df = pd.concat([stats_each_df, stats_df])
            
        index_to_drop = []
        
        for i in range(stats_each_df.shape[0]):
            if stats_each_df.iloc[i]['season'] not in ('22/23', '21/22'):
                index_to_drop.append(i)
        #stats_each_df =  stats_each_df.drop(index_to_drop)
        stats_each_df_dropped = stats_each_df.drop('season', axis = 1)
        
        stats = pd.DataFrame(columns = columns)
        
        stats = pd.concat([stats, stats_each_df_dropped.iloc[0].to_frame().T], ignore_index=True)
                
        for i in range(1, stats_each_df.shape[0]):
            if stats_each_df.iloc[i]['season'] in ('22/23', '21/22'):
                stats.iloc[0]['appearance'] += stats_each_df.iloc[i]['appearance']
                stats.iloc[0]['goals'] += stats_each_df.iloc[i]['goals']
                stats.iloc[0]['assists'] += stats_each_df.iloc[i]['assists']
                stats.iloc[0]['yellow cards'] += stats_each_df.iloc[i]['yellow cards']
                stats.iloc[0]['second yellow cards'] += stats_each_df.iloc[i]['second yellow cards']
                stats.iloc[0]['red cards'] += stats_each_df.iloc[i]['red cards']
                stats.iloc[0]['goals conceded'] += stats_each_df.iloc[i]['goals conceded']
                stats.iloc[0]['clean sheets'] += stats_each_df.iloc[i]['clean sheets']
                stats.iloc[0]['minutes played'] += stats_each_df.iloc[i]['minutes played']
       
        # Transform the data to per90
        per90 = stats.iloc[0]['minutes played']/90
        if per90 != 0:
            stats.iloc[0]['goals'] /= per90
            stats.iloc[0]['assists'] /= per90
            stats.iloc[0]['yellow cards'] /= per90
            stats.iloc[0]['second yellow cards'] /= per90
            stats.iloc[0]['red cards'] /= per90
            stats.iloc[0]['goals conceded'] /= per90
            stats.iloc[0]['clean sheets'] /= per90
        
        
        stats_player_df = pd.concat([stats_player_df, stats])
    
    return stats_player_df
            

# Sample_data là một list, mỗi phần tử là 1 dataframe tách ra từ sampled_players
sample_data = []
sample_data.append(sampled_players.iloc[:10])
i = 10
while i < sampled_players.shape[0]:
    if i >= sampled_players.shape[0] - 200:
        sample_data.append(sampled_players.iloc[i:])
    else:
        sample_data.append(sampled_players.iloc[i:i+200])
    i += 200 # mỗi sample chứa 200 dòng
print(len(sample_data))

i = 46
sample = obtain_players_stats(sample_data[i])
sample

file_name = "C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Player_stats/sample_data_" + str(i) + ".csv"
sample.to_csv(file_name) # Lưu thành file sample_data_i.csv

sample
