==> Source Code/Crawling/8_Crawl_market_value.ipynb <==
import pandas as pd
import numpy as np
import time
import re
from tqdm import tqdm_notebook
from torrequest import TorRequest
from bs4 import BeautifulSoup
from retrying import retry
from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.support.ui import WebDriverWait
from selenium.common.exceptions import TimeoutException
from selenium.webdriver.common.by import By
from selenium.webdriver.support import expected_conditions as ECimport 
import pandas as pd
import souper_v2
from tqdm import tqdm_notebook

sampled_players = pd.read_csv('C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Obtaining_players_and_teams/players_transfermarket.csv')
sampled_players

def get_soup(URL):
    
    chrome_options = Options()
    chrome_options.add_argument('--window-size=1,1')
    
    #options = webdriver.ChromeOptions()
    #options.add_argument('--disable-popup-blocking')
    
    driver = webdriver.Chrome(ChromeDriverManager().install(), options = chrome_options)
    wait = WebDriverWait(driver, 10)
    driver.get(URL)
    if len(driver.find_elements_by_css_selector("[name=ROBOTS]")) > 0:
        driver.get(URL)
    soup = BeautifulSoup(driver.page_source)
    driver.close()
    return soup

# Hàm đưa giá trị chuyển nhượng từ kiểu string -> numeric
# Ví dụ '€30.00m' -> '30000000'
def clean_value(s):
    if s[0] == '€':
        s = s[1:]
    s1 = s[:-1]
    #s1 = s1.results.map(lambda x: str(x).replace('.',''))
    s2 = float(s1)
    if s[len(s)-1] == 'm':
        s2 = s2*1000000
    elif s[len(s)-1] == 'k':
        s2 = s2*1000
    return s2

# Thu thập giá trị chuyển nhượng của từng cầu thủ
def obtain_market_value(players_df):
     
    links = players_df['link']
    nationality = ""
    
    returner = {'link': [], 'current_value': [], 'highest_value': []}
    
    index = 0
    
    for each_link in links:
        
        # Obtaining final link
        new_link = 'https://www.transfermarkt.co.uk' + each_link
        print(new_link)
        soup = get_soup(new_link)
        
        # Obtaining highest and current value
        #values = [x.text.strip() for x in soup.find_all('div', attrs={'class':'tm-player-market-value-development__current-and-max'})]
        
        try:
            a = soup.find('div', class_ = "tm-player-market-value-development__current-value")
            current_value = a.text.strip()
            print(current_value)
            current_value = clean_value(current_value)
            
        except:
            current_value = 0
            
        try:
            highest_value = soup.find('div', class_ = "tm-player-market-value-development__max-value").text
            print(highest_value)
            highest_value = clean_value(highest_value)
        except:
            highest_value = 0
                    
        returner['current_value'].append(current_value)
        returner['highest_value'].append(highest_value)
        returner['link'].append(each_link)
        
        print("Finish", index)
        index += 1
                    
    return returner

## Obtaining data by parts

# Sample_data là một list, mỗi phần tử là 1 dataframe tách ra từ sampled_players
sample_data = []
sample_data.append(sampled_players.iloc[:10])
i = 10
while i < sampled_players.shape[0]:
    if i >= sampled_players.shape[0] - 200:
        sample_data.append(sampled_players.iloc[i:])
    else:
        sample_data.append(sampled_players.iloc[i:i+200])
    i += 200 # mỗi sample chứa 200 dòng
print(len(sample_data))

i = 3
sample = obtain_market_value(sample_data[i])
sample

df = pd.DataFrame(sample)
df

file_name =  "C:/Users/Dell/Downloads/Work/TTDL/Đồ án/Data/Obtain_market_value/sample_data_" + str(i) + ".csv"
df.to_csv(file_name)

print(clean_value('10.00m'))
